Move tkeyHandler to separate file
